[Perpetual] move lotSize and lotPrecision to PerpetualQueriesContext
