+++ src/app/pages/PerpetualPage/components/RecentTradesTable/index.tsx
@@ -1,12 +1,12 @@
 import { AssetSymbolRenderer } from 'app/components/AssetSymbolRenderer';
 import { translations } from 'locales/i18n';
-import React, { useContext, useMemo } from 'react';
+import React, { useContext } from 'react';
 import { Trans } from 'react-i18next';
 import { PerpetualPair } from 'utils/models/perpetual-pair';
 import styles from './index.module.scss';
 import { RecentTradesTableRow } from './components/RecentTablesRow/index';
 import { RecentTradesContext } from './context';
-import { usePerpetual_queryPerpParameters } from '../../hooks/usePerpetual_queryPerpParameters';
+import { PerpetualQueriesContext } from '../../contexts/PerpetualQueriesContext';
 
 type RecentTradesTableProps = {
   pair: PerpetualPair;
@@ -16,13 +16,7 @@ export const RecentTradesTable: React.FC<RecentTradesTableProps> = ({
   pair,
 }) => {
   const { trades } = useContext(RecentTradesContext);
-  const perpParameters = usePerpetual_queryPerpParameters();
-  const lotPrecision = useMemo(() => {
-    const lotSize = Number(perpParameters.fLotSizeBC.toPrecision(8));
-    const lotPrecision = lotSize.toString().split(/[,.]/)[1]?.length || 0;
-
-    return lotPrecision;
-  }, [perpParameters.fLotSizeBC]);
+  const { lotPrecision } = useContext(PerpetualQueriesContext);
 
   return (
     <table className={styles.recentTradesTable}>
